requirements.txt
anthropic
python-dotenv 